Saves small memory, moves ship to initial position when player dies.

+++ arduboy-game/player.h
@@ -7,7 +7,7 @@
 struct Player {
   byte x;
   byte y;
-  int frame;
+  byte frame;
 
   void set() {
     x = 2;
